api/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
motor==3.3.2
boto3==1.34.1
redis[hiredis]==4.6.0
fastapi-cache2[redis]==0.2.1
